+++ ios/signer.h
@@ -85,5 +85,5 @@ const char* substrate_mini_secret_key_with_ref(struct ExternError*, int64_t seed
 const char* substrate_mini_secret_key(struct ExternError*, const char* suri);
 
 // qr fountain decoder
-const char * try_decode_qr_sequence(struct ExternError*, int64_t size, const char* data_json);
+const char * try_decode_qr_sequence(struct ExternError*, int size, const char* data_json);
